pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gh-audit"
dynamic = ["version"]
description = "GitHub audit, governance, and inventory for organizations"
readme = "README.md"
license = "BUSL-1.1"
requires-python = ">=3.11"
authors = [{ name = "N8 Group" }]
keywords = ["github", "audit", "governance", "inventory", "discovery"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Version Control :: Git",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "cryptography>=42.0.0",
    "httpx>=0.27.0",
    "jinja2>=3.1.0",
    "openpyxl>=3.1.0",
    "posthog>=3.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyjwt>=2.8.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "structlog>=24.0.0",
    "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "commitizen>=3.13.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "respx>=0.21.0",
    "ruff>=0.4.0",
]
build = ["pyinstaller>=6.0.0"]

[project.scripts]
gh-audit = "gh_audit.cli.app:run"

[project.urls]
Homepage = "https://n8-group.com"
Documentation = "https://github.com/n8group-oss/gh-audit"
Repository = "https://github.com/n8group-oss/gh-audit"
Issues = "https://github.com/n8group-oss/gh-audit/issues"

[tool.hatch.version]
path = "src/gh_audit/__about__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/gh_audit"]

[tool.hatch.envs.default]
features = ["dev"]

[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
cov = "pytest --cov=gh_audit --cov-report=term-missing {args:tests}"
lint = "ruff check src/ tests/"
lint-fix = "ruff check --fix src/ tests/"
format = "ruff format src/ tests/"
format-check = "ruff format --check src/ tests/"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = ["slow: marks tests as slow", "integration: marks integration tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
version_provider = "commitizen"
tag_format = "v$version"
version_scheme = "pep440"
version_files = [
    "src/gh_audit/__about__.py:__version__",
]
update_changelog_on_bump = true
changelog_file = "CHANGELOG.md"
changelog_incremental = true
